# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: CARVIS API
  version: 0.1.0
  description: |
    CARVIS Canvas Intelligence API — authentication, Canvas sync, voice commands,
    reminders, and the Chrome extension agent. All authenticated endpoints require
    a `jarvis_session` cookie or `x-session-token` header.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /canvas/sync:
    post:
      operationId: canvasSync
      tags: [canvas]
      summary: Sync courses, assignments, and grades from Canvas
      responses:
        "200":
          description: Sync completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  courseCount:
                    type: integer
        "400":
          description: Canvas not connected
        "500":
          description: Sync failed
...
  /canvas/assignments/toggle:
    patch:
      operationId: toggleAssignment
      tags: [canvas]
      summary: Mark an assignment as completed or incomplete
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assignmentId, completed]
              properties:
                assignmentId:
                  type: string
                completed:
                  type: boolean
      responses:
        "200":
          description: Updated
        "404":
          description: Assignment not found
  /canvas/activation:
    post:
      operationId: recordActivation
      tags: [canvas]
      summary: Record a one-time activation event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventType]
              properties:
                eventType:
                  type: string
                  enum: [first_sync_completed, first_question_asked, first_voice_used]
      responses:
        "200":
          description: Event recorded (or deduped)

  # ── Voice ─────────────────────────────────────────────────────────────────
...